# backend/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "Gather API",
    "description": "Social scheduling API for coordinating meetups with friends based on mutual availability"
  },
  "servers": [
    {
      "url": "https://api.gather.example.com",
      "description": "Production"
    },
    {
      "url": "http://localhost:3000",
      "description": "Local development"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Apple identity token"
      }
    },
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "enum": [
              false
            ]
          },
          "error": {
            "type": "string",
            "example": "Validation Error"
          },
          "message": {
            "type": "string",
            "example": "Invalid input provided"
          }
        },
        "required": [
          "success",
          "error",
          "message"
        ]
      },
...
      "CalendarListResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "enum": [
              true
            ]
          },
          "data": {
            "type": "object",
            "properties": {
              "connections": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/CalendarConnection"
                }
              }
            },
            "required": [
              "connections"
            ]
          }
        },
        "required": [
          "success",
          "data"
        ]
      },
...
      "SyncCalendarsResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "enum": [
              true
            ]
          },
          "data": {
            "type": "object",
            "properties": {
              "connections": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/CalendarConnection"
                }
              }
            },
            "required": [
              "connections"
            ]
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "success",
          "data"
        ]
      },
      "SyncCalendars": {
        "type": "object",
        "properties": {
          "calendars": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SyncCalendarEntry"
            }
          }
        },
        "required": [
          "calendars"
        ]
      },
...
    },
...
  },
  "paths": {
...
    "/calendars": {
      "get": {
        "tags": [
          "Calendars"
        ],
        "summary": "List calendar connections",
        "description": "Get all calendar connections for the current user",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of calendar connections",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CalendarListResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/calendars/sync": {
      "post": {
        "tags": [
          "Calendars"
        ],
        "summary": "Sync device calendars",
        "description": "Bulk sync calendars and their events from the device. Upserts calendar connections, syncs cached events, and removes deselected calendars.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SyncCalendars"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Calendars synced successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SyncCalendarsResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/calendars/google/auth-url": {
      "get": {
        "tags": [
          "Calendars",
          "Google"
        ],
        "summary": "Get Google OAuth URL",
        "description": "Get the Google OAuth consent URL for the current user to authorize calendar access",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "boolean",
              "nullable": true,
              "description": "Request export (write) scope in addition to read scopes"
            },
            "required": false,
            "description": "Request export (write) scope in addition to read scopes",
            "name": "includeExportScope",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "OAuth URL generated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GoogleAuthUrlResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/calendars/google/calendars": {
      "get": {
        "tags": [
          "Calendars",
          "Google"
        ],
        "summary": "List Google calendars",
        "description": "List the user's Google calendars (fetched live from Google API). Requires an existing Google connection.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of Google calendars",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GoogleCalendarListResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/calendars/google/select": {
      "post": {
        "tags": [
          "Calendars",
          "Google"
        ],
        "summary": "Select Google calendars to import",
        "description": "Choose which Google calendars to import for availability tracking",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GoogleSelectCalendars"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Calendar selection updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CalendarListResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/calendars/google/sync": {
      "post": {
        "tags": [
          "Calendars",
          "Google"
        ],
        "summary": "Sync Google calendars",
        "description": "Trigger a server-side re-sync of all connected Google calendars for the current user",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Google calendars synced",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SyncCalendarsResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/calendars/outlook/auth-url": {
      "get": {
        "tags": [
          "Calendars",
          "Outlook"
        ],
        "summary": "Get Outlook OAuth URL",
        "description": "Get the Outlook OAuth consent URL for the current user to authorize calendar access",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "boolean",
              "nullable": true,
              "description": "Request export (write) scope in addition to read scopes"
            },
            "required": false,
            "description": "Request export (write) scope in addition to read scopes",
            "name": "includeExportScope",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "OAuth URL generated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GoogleAuthUrlResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  }
}